Crate rtf_config

Config file parsing and validation for the Apollo Runtime Testing Framework.

Modules

checks

Helpers for checking config files

context

error

Generic helpers for working with operations that can return multiple errors, such as templating and checks.

formats

The various different config file formats that we support

providers

Providers are how we expose the rest of the framework to user facing config.

templating

Helpers for supporting minimal templating of user config files.

Macros

enum_impl_as_utf8_file_content

Helper macro for stamping out implementations of the AsUtf8FileContent trait on an enum where each variant is a wrapper around a type that already implements the trait.

enum_impl_check

Helper macro for stamping out implementations of the Check trait on an enum where each variant is a wrapper around a type that already implements the trait.

enum_impl_resolve_and_write

Helper macro for stamping out implementations of the ResolveAndWrite trait on an enum where each variant is a wrapper around a type that already implements the trait.

enum_impl_template

Helper macro for stamping out implementations of the Template trait on an enum where each variant is a wrapper around a type that already implements the trait.

impl_template

Helper for implementing the Template trait for types that only contain Fields at the top level. For types that contain collections of fields (e.g. Vecs and Maps) we need to write the implementation by hand.

Structs

ValueDefinition