rtf_config::context

Trait ResolutionContext

pub trait ResolutionContext {
    type PlatformClient: Client;
    type GithubClient: Client;
    type HttpClient: HttpClient;

    // Required methods
    fn http_client(&self) -> &Self::HttpClient;
    fn set_output_path(&mut self, path: impl Into<PathBuf>);
    fn output_path(&self) -> &Path;
    fn store_provider_output_path(
        &mut self,
        provider: Provider<'_>,
        path: PathBuf,
    );
    fn canonicalize_path(
        &self,
        relative_path: impl AsRef<Path>,
    ) -> Result<PathBuf>;
    fn path_kind(&self, path: impl AsRef<Path>) -> PathKind;
    fn read_path_to_string(&self, path: impl AsRef<Path>) -> Result<String>;
    fn write(
        &self,
        path: impl AsRef<Path>,
        content: impl AsRef<[u8]>,
    ) -> Result<()>;
    fn run_command_blocking<'a>(
        &self,
        prog: &str,
        args: impl IntoIterator<Item = &'a str>,
        env_vars: &HashMap<String, String>,
    ) -> Result<()>;
    fn remove_file(&self, path: impl AsRef<Path>) -> Result<()>;
    fn remove_dir_all(&self, path: impl AsRef<Path>) -> Result<()>;
    fn create_dir_all(&self, path: impl AsRef<Path>) -> Result<()>;

    // Provided methods
    fn platform_client(&self) -> Option<&Self::PlatformClient> { ... }
    fn github_client(&self) -> Option<&Self::GithubClient> { ... }
    fn known_provider_output_path(
        &self,
        provider: Provider<'_>,
    ) -> Option<PathBuf> { ... }
    async fn with_supergraph_details<T>(
        &self,
        graph_id: impl Into<String>,
        variant: impl Into<String>,
        f: impl FnOnce(&Arc<SupergraphDetails>) -> Result<T>,
    ) -> Result<T> { ... }
    fn dir_containing(&self, path: impl AsRef<Path>) -> PathBuf { ... }
    fn make_executable(&self, path: impl AsRef<Path>) -> Result<()> { ... }
    fn store_run_metadata(
        &mut self,
        _key: &'static str,
        _value: impl Into<String>,
    ) { ... }
    fn run_metadata(&self, _key: &str) -> Option<&str> { ... }

}

Types that implement ResolutionContext may be used to perform IO while checking and templating providers.

This trait is used to allow us to inject mock IO implementations in tests so we do not need to make use of temp directories or mock servers. For the most part the methods provided by this trait are wrappers around std::fs and API clients for making requests to third party APIs.

For the canonical real implementations that should be mocked see Context.

Required Associated Types

type PlatformClient: Client

type GithubClient: Client

type HttpClient: HttpClient

Required Methods

fn http_client(&self) -> &Self::HttpClient

fn set_output_path(&mut self, path: impl Into<PathBuf>)

fn output_path(&self) -> &Path

The top level output path being used.

fn store_provider_output_path(&mut self, provider: Provider<'_>, path: PathBuf)

Record the path that the given provider’s output was written to.

fn canonicalize_path(&self, relative_path: impl AsRef<Path>) -> Result<PathBuf>

Returns the canonical, absolute form of the path with all intermediate components normalized and symbolic links resolved.

fn path_kind(&self, path: impl AsRef<Path>) -> PathKind

Categorise the provided path.

If you cannot access the metadata of the file, e.g. because of a permission error or broken symbolic links, this will return PathKind::Missing.

fn read_path_to_string(&self, path: impl AsRef<Path>) -> Result<String>

Reads the entire contents of a file into a string.

Errors

This function will return an error if path does not already exist. Other errors may also be returned according to OpenOptions::open.

If the contents of the file are not valid UTF-8, then an error will also be returned.

While reading from the file, this function handles io::ErrorKind::Interrupted with automatic retries. See io::Read documentation for details.

fn write(&self, path: impl AsRef<Path>, content: impl AsRef<[u8]>) -> Result<()>

Writes a slice as the entire contents of a file.

This function will create a file if it does not exist, and will entirely replace its contents if it does.

Depending on the platform, this function may fail if the full directory path does not exist.

fn run_command_blocking<'a>( &self, prog: &str, args: impl IntoIterator<Item = &'a str>, env_vars: &HashMap<String, String>, ) -> Result<()>

Spawn the specified program as a subprocess with the provided arguments. This method will block until the process completes and return the stdout of the process as a utf-8 string.

fn remove_file(&self, path: impl AsRef<Path>) -> Result<()>

fn remove_dir_all(&self, path: impl AsRef<Path>) -> Result<()>

fn create_dir_all(&self, path: impl AsRef<Path>) -> Result<()>

Recursively create a directory and all of its parent components if they are missing.

If this function returns an error, some of the parent components might have been created already.

If the empty path is passed to this function, it always succeeds without creating any directories.

Provided Methods

fn platform_client(&self) -> Option<&Self::PlatformClient>

Provide a [Client][platform_query::Client] for making requests to the Apollo platform API.

If it is not possible for this current context to make requests to the platform API then this method should return None.

fn github_client(&self) -> Option<&Self::GithubClient>

Provide a [Client][github::Client] for making requests to the GitHub REST API.

If it is not possible for this current context to make requests to the GitHub API then this method should return None.

fn known_provider_output_path(&self, provider: Provider<'_>) -> Option<PathBuf>

Query the output path of a given provider.

async fn with_supergraph_details<T>( &self, graph_id: impl Into<String>, variant: impl Into<String>, f: impl FnOnce(&Arc<SupergraphDetails>) -> Result<T>, ) -> Result<T>

Make use of potentially cached supergraph details to derive related data. The default implementation of this method will simply pull the required [SupergraphDetails] and pass them to the mapping function provided. Custom implementations can be used to implement caching and other custom behaviour.

Panics

The default implementation will panic if ResolutionContext::platform_client is None.

fn dir_containing(&self, path: impl AsRef<Path>) -> PathBuf

fn make_executable(&self, path: impl AsRef<Path>) -> Result<()>

Changes the permissions of the specified file

fn store_run_metadata(&mut self, _key: &'static str, _value: impl Into<String>)

Store a key-value pair of run metadata for cross-phase communication.

fn run_metadata(&self, _key: &str) -> Option<&str>

Retrieve a previously stored run metadata value by key.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.