rtf_config

Module formats

The various different config file formats that we support

Structs

CustomProviderDefinition

Custom Provider Definition

EnvironmentConfig

Environment Config

Matrix

A matrix of user provided variable dimensions that is expanded out into multiple variable sets for templating the test plan containing the matrix.

RawTestPlanConfig

The raw format for parsing scenario config. This allows the ScenarioConfig and EnvironmentConfig to be retrieved from files or inline content before being fully templated and checked.

ScenarioConfig

Scenario Config

TestPlanConfig

The format for parsing scenario config

Enums

Error

Errors that can be encountered resolving config files

Type Aliases

Result