rtf_config::formats

Type Alias DockerComposeEnvironment

pub type DockerComposeEnvironment = ManifestEnvironment<ComposeResources>;

Aliased Type

pub struct DockerComposeEnvironment {
    pub resources: ComposeResources,
    pub file_providers: Vec<NamedFileProvider>,
    pub env_vars: HashMap<String, Field<Scalar>>,
    pub output_collection: OutputCollection,
}

Fields
resources: ComposeResourcesfile_providers: Vec<NamedFileProvider>

A list of all other files this environment depends on

env_vars: HashMap<String, Field<Scalar>>output_collection: OutputCollection
Implementations

impl DockerComposeEnvironment

pub fn file_provider_services( &self, ctx: &impl ResolutionContext, ) -> Result<FileProviderServices>

Categorise docker compose services based on how they accesses file provider output.

pub fn services(&self) -> Result<Vec<EnvironmentService>>

The services this environment will deploy, read out of its compose file contents.

Requires that only inline providers are present, erroring if any non-inline providers are encountered. inline_manifests can be used to enforce this invariant.