rtf_config

Macro enum_impl_resolve_and_write

macro_rules! enum_impl_resolve_and_write {
    ($enum:ident => $($variant:ident),+) => { ... };
}

Helper macro for stamping out implementations of the ResolveAndWrite trait on an enum where each variant is a wrapper around a type that already implements the trait.