rtf_config::providers

Module file

The core FileProvider trait and currently supported file provider implementations.

Modules

apollo

custom

github

utility

File providers that act as combinators or otherwise modify the output of other providers

Structs

InlineFile

Inline file

NamedFileProvider

Named File Provider

RelativeFile

Relative path

RequiredFile

Required file

Enums

FileProvider

File Provider

RawSource

Config Source

SourceDir

The source directory of where a particular config file was obtained.