rtf_config::run

Trait Execute

pub trait Execute: RunProviders {
    // Required methods
    fn command_name(&self) -> &str;
    fn execute(
        &self,
        out_dir: &Path,
        output_path: &Path,
        ctx: &impl ResolutionContext,
    ) -> Result<()>;

    // Provided methods
    async fn run_providers_and_execute(
        &self,
        name: &str,
        out_dir: &Path,
        output_path: PathBuf,
        providers_dir: PathBuf,
        ctx: &mut impl ResolutionContext,
    ) -> Result<PathBuf> { ... }
    async fn run_providers_and_execute_for_output(
        &self,
        name: &str,
        out_dir: &Path,
        ctx: &mut impl ResolutionContext,
    ) -> Result<String> { ... }
}

Required Methods

fn command_name(&self) -> &str

fn execute( &self, out_dir: &Path, output_path: &Path, ctx: &impl ResolutionContext, ) -> Result<()>

Execute this command with the specified environment, returning the output path used.

RunProviders::run_providers must have been run successfully before calling this method in order to ensure that all file providers have written out their file content to the expected location.

Provided Methods

async fn run_providers_and_execute( &self, name: &str, out_dir: &Path, output_path: PathBuf, providers_dir: PathBuf, ctx: &mut impl ResolutionContext, ) -> Result<PathBuf>

Run all of the FileProviders associated with this command and write out their file contents to the specified directory before executing the command with the specified environment, returning the the output path passed to the command.

async fn run_providers_and_execute_for_output( &self, name: &str, out_dir: &Path, ctx: &mut impl ResolutionContext, ) -> Result<String>

Run all of the FileProviders associated with this command and write out their file contents to the specified directory before executing the command with the specified environment, returning the the content written to OUTPUT_PATH and removing the output file if it was created.

This method assumes that the command being executed is writing a single file out at the provided output path rather than a directory of files. To run a command and leave the resources available on disk after execution, use run_providers_and_execute instead.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.