rtf_config

Module formats

The various different config file formats that we support

Structs

CustomProviderDeclaration

Custom Provider Declaration

CustomProviderDefinition

Custom Provider Definition

DockerCommand

DockerComposeEnvironment

DockerScenario

EnvironmentConfig

Environment Config

FileProviderServices

Categorisation of compose services based on how they access file provider output.

Generic

Marker for the generic (multi-execution-type) test plan variant.

Matrix

A matrix of user provided variable dimensions that is expanded out into multiple variable sets for templating the test plan containing the matrix.

OutputCollection

Defines the data that will be collected when a test execution has completed

PrometheusQuery

RawTestPlanConfig

The raw format for parsing scenario config. This allows the ScenarioConfig and EnvironmentConfig to be retrieved from files or inline content before being fully templated and checked.

ScenarioConfig

Scenario Config

ScriptEnvironment

Sources

The source paths of each of the configs for a given test plan.

TestPlan

The format for parsing a test plan config

Enums

EnvironmentExecution

Error

Errors that can be encountered resolving config files

ScenarioExecution

Traits

Execution

Associates a concrete (Scenario, Environment) execution type pair.

Type Aliases

Result

TestPlanConfig

Test plan that accepts any scenario/environment execution type.