rtf_config::templating

Trait Template

pub trait Template {
    // Required methods
    fn required_variables(&self) -> Vec<String>;
    fn validate_context(
        &self,
        path: &mut Vec<String>,
        allowed_variables: &HashSet<&String>,
        file_source: &StableSource,
        ctx: &TemplateContext,
    ) -> Result<()>;
    fn try_template(
        &mut self,
        path: &mut Vec<String>,
        file_source: &StableSource,
        ctx: &TemplateContext,
    ) -> Result<()>;

    // Provided methods
    fn validate_context_nested(
        &self,
        path: &mut Vec<String>,
        tail: &str,
        allowed_variables: &HashSet<&String>,
        file_source: &StableSource,
        ctx: &TemplateContext,
    ) -> Result<()> { ... }
    fn try_template_nested(
        &mut self,
        path: &mut Vec<String>,
        tail: &str,
        file_source: &StableSource,
        ctx: &TemplateContext,
    ) -> Result<()> { ... }
    fn try_template_known(
        &mut self,
        path: &mut Vec<String>,
        file_source: &StableSource,
        ctx: &TemplateContext,
    ) -> Result<Vec<String>> { ... }
}

In order to support controlled templating of config files with Scalar variables we make use of a wrapper Field type to identify where variables need to be injected. A type that implements Template supports walking its contents to locate and template fields using a provided map of scalar variables.

Required Methods

fn required_variables(&self) -> Vec<String>

The list of template variables that are required to template this type fully.

fn validate_context( &self, path: &mut Vec<String>, allowed_variables: &HashSet<&String>, file_source: &StableSource, ctx: &TemplateContext, ) -> Result<()>

Check that the provided TemplateContext is sufficient for templating all fields and custom providers under this type.

fn try_template( &mut self, path: &mut Vec<String>, file_source: &StableSource, ctx: &TemplateContext, ) -> Result<()>

Attempt to resolve all pending Fields, appending encountered errors to the errs vec provided.

Provided Methods

fn validate_context_nested( &self, path: &mut Vec<String>, tail: &str, allowed_variables: &HashSet<&String>, file_source: &StableSource, ctx: &TemplateContext, ) -> Result<()>

fn try_template_nested( &mut self, path: &mut Vec<String>, tail: &str, file_source: &StableSource, ctx: &TemplateContext, ) -> Result<()>

Attempt to resolve all pending Fields when this type is a child of some parent Template, appending encountered errors to the errs vec provided. The provided tail will be appended to path before calling through to Template::try_template.

fn try_template_known( &mut self, path: &mut Vec<String>, file_source: &StableSource, ctx: &TemplateContext, ) -> Result<Vec<String>>

Attempt to resolve all known Fields, reporting required variables that are not present in the provided map. If there are any deserialization errors then then this method as an aggregate operation will fail.