rtf_config::templating

Trait Template

pub trait Template {
    // Required methods
    fn has_pending_fields(&self) -> bool;
    fn required_values(&self) -> Vec<String>;
    fn try_template(
        &mut self,
        path: &mut Vec<String>,
        values: &HashMap<String, Scalar>,
    ) -> Result<()>;

    // Provided methods
    fn try_template_nested(
        &mut self,
        path: &mut Vec<String>,
        tail: &str,
        values: &HashMap<String, Scalar>,
    ) -> Result<()> { ... }
    fn try_template_known(
        &mut self,
        values: &HashMap<String, Scalar>,
    ) -> Result<Vec<String>> { ... }
}

In order to support controlled templating of config files with Scalar values we make use of a wrapper Field type to identify where values need to be injected. A type that implements Template supports walking its contents to locate and template fields using a provided map of scalar values.

Required Methods

fn has_pending_fields(&self) -> bool

Whether or not there are any pending Fields contained within this value.

fn required_values(&self) -> Vec<String>

The list of template values that are required to template this type fully.

fn try_template( &mut self, path: &mut Vec<String>, values: &HashMap<String, Scalar>, ) -> Result<()>

Attempt to resolve all pending Fields, appending encountered errors to the errs vec provided.

Provided Methods

fn try_template_nested( &mut self, path: &mut Vec<String>, tail: &str, values: &HashMap<String, Scalar>, ) -> Result<()>

Attempt to resolve all pending Fields when this type is a child of some parent Template, appending encountered errors to the errs vec provided. The provided tail will be appended to path before calling through to Template::try_template.

fn try_template_known( &mut self, values: &HashMap<String, Scalar>, ) -> Result<Vec<String>>

Attempt to resolve all known Fields, reporting required values that are not present in the provided map. If there are any deserialization errors then then this method as an aggregate operation will fail.